+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Added missing include for std functions
 - Updated CMake configuration adding install target
+- Added run_tests target to run all unit tests
 - First release:
   - Support for water and steam temperature sensor
   - Support for serial interface